+++ plugins/a11y/CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.1.0
+- Added new reporter file to help generating of landing page 
+- Update webpack plugin hook processes
+- Updated npm packages
+- Added @caweb/template peerDependency
+
 v1.0.9
 - Update npm packages
 - Added npm devDependencies packages